# docs/swagger.yaml
basePath: /
definitions:
...
  github_com_NeuralTrust_TrustGate_pkg_api_handler_http_helpers.ErrorBody:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
...
  github_com_NeuralTrust_TrustGate_pkg_api_handler_http_registry_request.APIKeyAuthRequest:
    properties:
      api_key:
        description: '#nosec G117'
        type: string
      header_name:
        type: string
      header_value:
        type: string
      param_location:
        type: string
      param_name:
        type: string
      param_value:
        type: string
    type: object
  github_com_NeuralTrust_TrustGate_pkg_api_handler_http_registry_request.AWSAuthRequest:
    properties:
      access_key_id:
        type: string
      region:
        type: string
      role:
        type: string
      secret_access_key:
        description: '#nosec G117'
        type: string
      session_token:
        description: '#nosec G117'
        type: string
      use_role:
        type: boolean
    type: object
  github_com_NeuralTrust_TrustGate_pkg_api_handler_http_registry_request.AzureAuthRequest:
    properties:
      api_key:
        description: '#nosec G117'
        type: string
      client_id:
        type: string
      client_secret:
        description: '#nosec G117'
        type: string
      endpoint:
        type: string
      tenant_id:
        type: string
      use_managed_identity:
        type: boolean
      version:
        type: string
    type: object
  github_com_NeuralTrust_TrustGate_pkg_api_handler_http_registry_request.CreateRegistryRequest:
    properties:
      auth:
        $ref: '#/definitions/github_com_NeuralTrust_TrustGate_pkg_api_handler_http_registry_request.TargetAuthRequest'
      description:
        type: string
      enabled:
        type: boolean
      health_checks:
        $ref: '#/definitions/github_com_NeuralTrust_TrustGate_pkg_api_handler_http_registry_request.HealthChecksRequest'
      mcp_target:
        $ref: '#/definitions/github_com_NeuralTrust_TrustGate_pkg_api_handler_http_registry_request.MCPTargetRequest'
      name:
        type: string
      provider:
        type: string
      provider_options:
        additionalProperties: {}
        type: object
      type:
        type: string
    type: object
  github_com_NeuralTrust_TrustGate_pkg_api_handler_http_registry_request.HealthChecksRequest:
    properties:
      headers:
        additionalProperties:
          type: string
        type: object
      interval:
        type: integer
      passive:
        type: boolean
      path:
        type: string
      threshold:
        type: integer
    type: object
  github_com_NeuralTrust_TrustGate_pkg_api_handler_http_registry_request.MCPAuthRequest:
    properties:
      actor:
        type: string
      audience:
        type: string
      authorize_url:
        type: string
      client_id:
        type: string
      client_secret:
        description: '#nosec G117'
        type: string
      expected_audience:
        type: string
      header:
        type: string
      mode:
        type: string
      pattern:
        type: string
      provider:
        type: string
      registration:
        type: string
      resource:
        type: string
      scope:
        type: string
      scopes:
        items:
          type: string
        type: array
      token_url:
        type: string
      value:
        description: '#nosec G117'
        type: string
    type: object
  github_com_NeuralTrust_TrustGate_pkg_api_handler_http_registry_request.MCPTargetRequest:
    properties:
      auth:
        $ref: '#/definitions/github_com_NeuralTrust_TrustGate_pkg_api_handler_http_registry_request.MCPAuthRequest'
      code:
        type: string
      headers:
        additionalProperties:
          type: string
        type: object
      transport:
        type: string
      url:
        type: string
    type: object
  github_com_NeuralTrust_TrustGate_pkg_api_handler_http_registry_request.TargetAuthRequest:
    properties:
      api_key:
        $ref: '#/definitions/github_com_NeuralTrust_TrustGate_pkg_api_handler_http_registry_request.APIKeyAuthRequest'
      aws:
        $ref: '#/definitions/github_com_NeuralTrust_TrustGate_pkg_api_handler_http_registry_request.AWSAuthRequest'
      azure:
        $ref: '#/definitions/github_com_NeuralTrust_TrustGate_pkg_api_handler_http_registry_request.AzureAuthRequest'
      gcp_service_account:
        type: string
      oauth:
        $ref: '#/definitions/github_com_NeuralTrust_TrustGate_pkg_api_handler_http_registry_request.TargetOAuthConfigRequest'
      type:
        type: string
    type: object
  github_com_NeuralTrust_TrustGate_pkg_api_handler_http_registry_request.TargetOAuthConfigRequest:
    properties:
      audience:
        type: string
      client_id:
        type: string
      client_secret:
        description: '#nosec G117'
        type: string
      code:
        type: string
      code_verifier:
        type: string
      extra:
        additionalProperties:
          type: string
        type: object
      grant_type:
        type: string
      password:
        description: '#nosec G117'
        type: string
      redirect_uri:
        type: string
      refresh_token:
        description: '#nosec G117'
        type: string
      scopes:
        items:
          type: string
        type: array
      token_url:
        type: string
      use_basic_auth:
        type: boolean
      username:
        type: string
    type: object
...
  github_com_NeuralTrust_TrustGate_pkg_api_handler_http_registry_response.APIKeyAuthResponse:
    properties:
      api_key:
        description: '#nosec G117'
        type: string
      header_name:
        type: string
      header_value:
        type: string
      param_location:
        type: string
      param_name:
        type: string
      param_value:
        type: string
    type: object
  github_com_NeuralTrust_TrustGate_pkg_api_handler_http_registry_response.AWSAuthResponse:
    properties:
      access_key_id:
        type: string
      region:
        type: string
      role:
        type: string
      secret_access_key:
        description: '#nosec G117'
        type: string
      session_token:
        description: '#nosec G117'
        type: string
      use_role:
        type: boolean
    type: object
  github_com_NeuralTrust_TrustGate_pkg_api_handler_http_registry_response.AzureAuthResponse:
    properties:
      api_key:
        description: '#nosec G117'
        type: string
      client_id:
        type: string
      client_secret:
        description: '#nosec G117'
        type: string
      endpoint:
        type: string
      tenant_id:
        type: string
      use_managed_identity:
        type: boolean
      version:
        type: string
    type: object
  github_com_NeuralTrust_TrustGate_pkg_api_handler_http_registry_response.HealthChecksResponse:
    properties:
      headers:
        additionalProperties:
          type: string
        type: object
      interval:
        type: integer
      passive:
        type: boolean
      path:
        type: string
      threshold:
        type: integer
    type: object
  github_com_NeuralTrust_TrustGate_pkg_api_handler_http_registry_response.ListRegistryResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/github_com_NeuralTrust_TrustGate_pkg_api_handler_http_registry_response.RegistryResponse'
        type: array
      page:
        type: integer
      size:
        type: integer
      total:
        type: integer
    type: object
  github_com_NeuralTrust_TrustGate_pkg_api_handler_http_registry_response.MCPAuthResponse:
    properties:
      actor:
        type: string
      audience:
        type: string
      authorize_url:
        type: string
      client_id:
        type: string
      client_secret:
        description: '#nosec G117 -- masked before serialization'
        type: string
      expected_audience:
        type: string
      header:
        type: string
      mode:
        type: string
      pattern:
        type: string
      provider:
        type: string
      registration:
        type: string
      resource:
        type: string
      scope:
        type: string
      scopes:
        items:
          type: string
        type: array
      token_url:
        type: string
      value:
        description: '#nosec G117 -- masked before serialization'
        type: string
    type: object
  github_com_NeuralTrust_TrustGate_pkg_api_handler_http_registry_response.MCPTargetResponse:
    properties:
      auth:
        $ref: '#/definitions/github_com_NeuralTrust_TrustGate_pkg_api_handler_http_registry_response.MCPAuthResponse'
      code:
        type: string
      headers:
        additionalProperties:
          type: string
        type: object
      transport:
        type: string
      url:
        type: string
    type: object
  github_com_NeuralTrust_TrustGate_pkg_api_handler_http_registry_response.RegistryResponse:
    properties:
      auth:
        $ref: '#/definitions/github_com_NeuralTrust_TrustGate_pkg_api_handler_http_registry_response.TargetAuthResponse'
      created_at:
        type: string
      description:
        type: string
      enabled:
        type: boolean
      gateway_id:
        type: string
      health_checks:
        $ref: '#/definitions/github_com_NeuralTrust_TrustGate_pkg_api_handler_http_registry_response.HealthChecksResponse'
      id:
        type: string
      mcp_target:
        $ref: '#/definitions/github_com_NeuralTrust_TrustGate_pkg_api_handler_http_registry_response.MCPTargetResponse'
      name:
        type: string
      provider:
        type: string
      provider_options:
        additionalProperties: {}
        type: object
      type:
        type: string
      updated_at:
        type: string
    type: object
  github_com_NeuralTrust_TrustGate_pkg_api_handler_http_registry_response.TargetAuthResponse:
    properties:
      api_key:
        $ref: '#/definitions/github_com_NeuralTrust_TrustGate_pkg_api_handler_http_registry_response.APIKeyAuthResponse'
      aws:
        $ref: '#/definitions/github_com_NeuralTrust_TrustGate_pkg_api_handler_http_registry_response.AWSAuthResponse'
      azure:
        $ref: '#/definitions/github_com_NeuralTrust_TrustGate_pkg_api_handler_http_registry_response.AzureAuthResponse'
      gcp_service_account:
        type: string
      oauth:
        $ref: '#/definitions/github_com_NeuralTrust_TrustGate_pkg_api_handler_http_registry_response.TargetOAuthConfigResponse'
      type:
        type: string
    type: object
  github_com_NeuralTrust_TrustGate_pkg_api_handler_http_registry_response.TargetOAuthConfigResponse:
    properties:
      audience:
        type: string
      client_id:
        type: string
      client_secret:
        description: '#nosec G117'
        type: string
      extra:
        additionalProperties:
          type: string
        type: object
      grant_type:
        type: string
      refresh_token:
        description: '#nosec G117'
        type: string
      scopes:
        items:
          type: string
        type: array
      token_url:
        type: string
      use_basic_auth:
        type: boolean
    type: object
...
info:
  contact:
    email: support@neuraltrust.ai
    name: NeuralTrust
    url: https://neuraltrust.ai/contact
  description: Administrative API for managing gateways and their registries, policies,
    consumers, roles and auth credentials.
  title: TrustGate Admin API
  version: "1.0"
paths:
...
  /{consumer_slug}/v1/chat/completions:
    post:
      consumes:
      - application/json
      description: 'Forwards an OpenAI Chat Completions request to the selected provider.
        Proxy plane route: /{consumer_slug}/v1/chat/completions. Other fixed routes
        include /v1/messages (Anthropic) and /v1/responses (OpenAI Responses).'
      parameters:
      - description: Consumer slug
        in: path
        name: consumer_slug
        required: true
        type: string
      - description: API key for inline consumers
        in: header
        name: X-AG-API-Key
        type: string
      - description: Bearer token for OAuth2 or OIDC consumers
        in: header
        name: Authorization
        type: string
      - description: Gateway slug when using header-based gateway discovery
        in: header
        name: X-AG-Gateway-Slug
        type: string
      - description: OpenAI Chat Completions request body
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_NeuralTrust_TrustGate_pkg_api_handler_http_helpers.ErrorBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_NeuralTrust_TrustGate_pkg_api_handler_http_helpers.ErrorBody'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_NeuralTrust_TrustGate_pkg_api_handler_http_helpers.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_NeuralTrust_TrustGate_pkg_api_handler_http_helpers.ErrorBody'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/github_com_NeuralTrust_TrustGate_pkg_api_handler_http_helpers.ErrorBody'
      summary: Proxy chat completion
      tags:
      - proxy
...
  /v1/gateways/{gateway_id}/registries:
    get:
      description: Returns a paginated list of registries in a gateway.
      parameters:
      - description: Gateway id
        format: uuid
        in: path
        name: gateway_id
        required: true
        type: string
      - description: Filter by name (substring match)
        in: query
        name: name
        type: string
      - description: Page number (1-based)
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_NeuralTrust_TrustGate_pkg_api_handler_http_registry_response.ListRegistryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_NeuralTrust_TrustGate_pkg_api_handler_http_helpers.ErrorBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_NeuralTrust_TrustGate_pkg_api_handler_http_helpers.ErrorBody'
      security:
      - BearerAuth: []
      summary: List registries
      tags:
      - registries
    post:
      consumes:
      - application/json
      description: Creates a new backend in a gateway.
      parameters:
      - description: Gateway id
        format: uuid
        in: path
        name: gateway_id
        required: true
        type: string
      - description: Registry to create
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_NeuralTrust_TrustGate_pkg_api_handler_http_registry_request.CreateRegistryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_NeuralTrust_TrustGate_pkg_api_handler_http_registry_response.RegistryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_NeuralTrust_TrustGate_pkg_api_handler_http_helpers.ErrorBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_NeuralTrust_TrustGate_pkg_api_handler_http_helpers.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_NeuralTrust_TrustGate_pkg_api_handler_http_helpers.ErrorBody'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_NeuralTrust_TrustGate_pkg_api_handler_http_helpers.ErrorBody'
      security:
      - BearerAuth: []
      summary: Create a backend
      tags:
      - registries
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"